chore: adjust onboarding experience - adjust unit tests
