+++ apps/trading/components/market-selector/index.ts
@@ -1,6 +1,3 @@
 export * from './market-selector';
 export * from './market-selector-item';
 export * from './market-selector-button';
-export { Product } from './product-selector';
-export { Sort } from './sort-dropdown';
-export { useMarketSelectorList } from './use-market-selector-list';
