+++ apps/trading/components/welcome-dialog/welcome-dialog-content.tsx
@@ -6,21 +6,27 @@ import { Links, Routes } from '../../pages/client-router';
 import { Networks, useEnvironment } from '@vegaprotocol/environment';
 import type { ReactNode } from 'react';
 import { useOnboardingStore } from './welcome-dialog';
-import { useMarketSelectorList, Product, Sort } from '../market-selector';
-
-const FILTERS = {
-  searchTerm: '',
-  product: Product.All,
-  sort: Sort.TopTraded,
-  assets: [],
-};
+import { useMarketList } from '@vegaprotocol/markets';
+import { isMarketActive } from '../../lib/utils';
+import orderBy from 'lodash/orderBy';
+import { priceChangePercentage } from '@vegaprotocol/utils';
 
 export const WelcomeDialogContent = () => {
   const { VEGA_ENV } = useEnvironment();
 
   const dismiss = useOnboardingStore((store) => store.dismiss);
   const navigate = useNavigate();
-  const { markets } = useMarketSelectorList(FILTERS);
+  const { data } = useMarketList();
+  const markets = orderBy(
+    data?.filter((m) => isMarketActive(m.state)) || [],
+    [
+      (m) => {
+        if (!m.candles?.length) return 0;
+        return Number(priceChangePercentage(m.candles.map((c) => c.close)));
+      },
+    ],
+    ['desc']
+  );
   const explore = () => {
     const marketId = markets?.[0].id ?? '';
     const link = marketId
